restructuring backend project
